+++ src/main/java/core/basesyntax/figure/Rectangle.java
@@ -1,28 +1,28 @@
 package core.basesyntax.figure;
 
 public class Rectangle extends Figure {
-    private double width;
-    private double length;
+    private int width;
+    private int length;
 
-    public Rectangle(String color, double width, double length) {
+    public Rectangle(String color, int width, int length) {
         super(color);
         this.width = width;
         this.length = length;
     }
 
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
 
-    public void setWidth(double width) {
+    public void setWidth(int width) {
         this.width = width;
     }
 
-    public double getLength() {
+    public int getLength() {
         return length;
     }
 
-    public void setLength(double length) {
+    public void setLength(int length) {
         this.length = length;
     }
